libOverlay: Added window calculation logic for interlaced panels
Description# In case of HDTV in interlaced mode, the timings for height
would be half of the progressive mode. The current window calculation
logic is for progressive panels, and when it is applied for Interlaced
panels, this results in downscaling the width to half.
This is fixed by taking the TV CEA code into account while
arriving at the final window settings.

Validation# Tested AV playback from Gallery on TV in
progressive mode and interlaced mode.

DRT# OMAPS00233143 - [HDMI compliance] Wrong aspect ratio in interlace mode

patchset3# updated the display code parsing logic as per the new
implementation in kernel

Change-Id: Id418cb85ae893b503e5b1dfa4f221112c83f4714
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>