libOverlay: sysfile read/write permissions are modified
Description# Currently we are opening the file in O_RDWR
irrespective of whether it is used for read or write.
This one would fail unless both read & write permissions
are given for all the sysFS entries.
This is fixed by opening the file with the correct
permissions.

Validation# AV playback on LCD and HDMI.

Signed-off-by: Sunita Nadampalli <sunitan@ti.com>

Patch Set 2: Simple cherry-pick, no changes made.  Taking
approvals, verifications and build test from patch set 1.
Signed-off-by: Jean Johnson <jean-johnson@ti.com>
Change-Id: Ia29b10cad1918f18968924bd2123f174743f7fb7