OMAP4 CameraHAL: Update reference count correctly if frame is dropped in context of sendFrame
This scenario was discovered in slow motion video encode use-case when the OverlayDisplayAdapter
drops the frames to render 120 fps stream from camera at 30 fps - since the reference count
for video encoder frames is incremented after sendFrame for display is called and display
returns the frame in sendFrame call context itself, we see the reference count for the frame
becoming zero and the frame queued back to ducati before even sending the frame to the encoder.
This patch ensures that the reference counts for the frame is updated in advance before sending
the callbacks and ensures that the reference count is checked correctly.

Signed-off-by: Sundar Raman <sunds@ti.com>

Patch set 3:  Breaking dependency on patch 12112.  Fixed simple merge conflict.
Taking approvals from patch set 2.
Signed-off-by: Jean Johnson <jean-johnson@ti.com>
Change-Id: Ib5df3f709decdd0477aa51bc9158e79e8a8ee872