CameraHAL OMAP4 : Support for Camera Orientation
This patch adds support to rotate the preview buffers.
This is done as part of Video Record Rotation.
1. The width and height are swapped for 90 and 270 degree rotation
by the application. This is taken care while validation of parameters
2. Based on Orientation angle is provided by the application.
3. Based on this Camera is counter rotated.
Patch Set #3: Fixes DR: OMAPS00233603

As per design in Ducati, for resolution with height > 1008,resolution
cannot be set without configuring orientation. So the current solution
sets a intermediate resolution, sets orientation and then the requested
resolution when the component is in loaded state.
Application/ end user are not affected by this change
Patch set#5 : Addressed review comments
Patch set#6 : No Changes. Edited commit message
Change-Id: Ie65080c6c2ac4d7851afe227e65148a1da38c9ad
Signed-off-by: Sathya Prakash M R <a0393677@ti.com>