libOverlay zOrder static assignment for HDMI usecases
Description# Recently libOverlay is modified to dynamically calculate the
zOrder for the video pipes based on the all active video pipes.
 This logic works fine if all the video pipes are controlled via libOverlay.
But for UI cloning on HDMI, video1 pipe is used in system server
UI cloning service, where video1 zorder is always zet to 0.
So, with the dynamic assigmnet there is a clash of video1 and the second
video pipe used for Video playback on HDTV.
This is fixed by statically assigning the zorder for video pipe, if the
manager if TV.And one more reason for this status assigment is, this sequence
is required to make rotation usecases work.

Validation: AV playback + UI cloning on HDTV and LCD, and Camera preview on
LCD and HDMI.

Change-Id: If1ae06b10115ea2f00f84f32d1e1faf286601d86
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>