OMAP4: AudioHAL: Modem: handling non-supported devices
If audio voice call is routed to a non-supported devices
instead to stop the audio modem HAL, the previous audio
device will be used or if no previous device
the HANDSET device will be used.

Example of problem:

If audio multimiedia is played over HDMI, when the MO
voice call starts the first device used is HDMI
(not supported in voice call in gingerbread)
even just after it is routed to handset.

Change-Id: I427353e9e0d425a80e1ffcc7d21748a1cee29b78