TIHardware renderer changes for dynamic resize
Description# TIHardware Renderer changes to implement
OMXRenderer interface for resize. The changes include
Video W&H update on resize along with the decoded buffer
width & height.

DRT#OMAPS00236082 - Multiple robustness test clips are playing with noise
in the right side of screen due to incorrect port reconfig handling
Note#This is one of the multiple patches for this DR
This patch is dependent on
http://review.omapzoom.org/#change,12704

Change-Id: I039834f64d6c54238a8f1a4f8fe059cdaebaa25d
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>