OMAP4: AudioHAL: Modem: add auto sample rate feature
when auto sample rate feature is enable, the voice call
sample rate used during the call is not fixed.
It depends on if the network start the voice call
in AMR-NB (8KHz) or AMR-WB(16KHz).
The aim is to use the most appropriate sample rate to
reduce the modem power consuption by avoiding enabling
modem SRC and acoustics chain not adapted to the network
vocoder used.

Change-Id: I52cc6e84fbab3026fdbdb4a484aad9bce75e00af