OMAP4 AUDIO HAL: Capture-Playback Volume for DL1 not being re-muted after FM Rx use case.
When taking video immediately after the FM Rx use case, the device plays back
the audio it is recording for the A/V capture.  This mutes the playback stream,
as it is being unmuted for the FM case.  This mixer control does not get reset for
all non-FM use cases.

Change-Id: I59751e2108e7a752c83374440766d780de3af719
Signed-off-by: Ricky Bhatia <a0866433@ti.com>