CameraHAL OMAP4: Adds a timeout during semaphore wait in AppCallbackNotifier
- This avoids a possible deadlock that could occur when smooth zoom
is running and the client tries to stop the preview. In this exact
moment the AppCallbackNotifier might be calling the client in order
to notify him on the next zoom step and the client might have
already called stopPreview() and unable to process the ZoomCallback.
On the other hand stopPreview will try to stop AppCallbackNotifier
which is not possible and will lead to a deadlock.

Signed-off-by: Emilian Peev <epeev@mm-sol.com>

patch set#2 Rebased.

Change-Id: I2bf2370908768768225abd7d2e11c8c731ae298a
Signed-off-by: Solaiyappan Saravanan <saravanan.s@ti.com>