OMAP4: ALSA: Add flexibility in DL1/DL2 mono choice
In Phoenix (TWL6040) audio codec,
headset and earpiece are connected to DL1 and
aux and speaker are connected to DL2.

Earpiece needs to be mono as it used only right sample of DL1.
Headset can be mono or stereo independently of earpiece.
Aux can be mono or stereo independently of speaker
Speaker can be mono or stereo independently of aux.

Change-Id: Iab74ec4764fc0b4616987fc340a5941a5b3ae910