LibOverlay changes to remove unnecessary checks for rotation
Description# LibOverlay commit method is modified to remove
the condition: "If the Overlay rotation is changed, without
actually changing the window size" ignore the rotation setting
during this commit call. This can happen in two scenarios:
(1) during every rotation where rotation angle is commited first
 and then the new window size, and
(2) During the rotation from 90 to 270 degrees clock wise or anti-clockwise.
In this scenario, the window size remains the same, but rotation
degree alone changes.
Because of the check in overlay, during the first scenario, the
rotation happens during the subsequent window change and rotation is smooth.
But in case of second scenario, the rotation would never happen, hence
results in inverted video.

DRT#OMAPS00238815 - PT AD HOC-  Progressing Bar is not rotating
ad hoc with the AV playback rotation

Change-Id: I2fc0fbf8d13f200e67c9907195b10b1c8a87cd6e
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>