Thermal Manager: Updated management of ALERT zone
If the OMAP junction temperature reaches the ALERT zone, then
the maximum CPU frequency could be limited if the current frequency is not
the higher possible one.
A SW race condition may be created between the thermal event and the
reading of current frequency. Finally, the scaling_max_freq could be
set to a low value.
To avoid this potential race condition, the scaling_max_freq is set to the next
lowest frequency instead of the current frequency.

Change-Id: Icc038e2b95b62dc54acbecb896b8f1d0adf1169d
Signed-off-by: Sebastien Sabatier <s-sabatier1@ti.com>