Thermal Manager: Introduce the PCB temp sensor manager for 4430
This patch changes the pcb temp manager to adjust the OMAP4430
duty cycle parameters. The change is based on duty cycle sections.

Here are a short list of features:
. The user must be capable to configure the duty cycle parameters files
. The user must be capable to use a sensor which generates thermal events.
	(but not able to test)
. The user must be capable of setting The policy to react on pcb temp
alerts and readjust the duty cycle and the sensor thresholds
(but not able to test)
. The user must be capable of setting the policy to poll the pcb
sensor, instead of reacting on temp alerts.
. The user must be capable to choose between polling strategy and
event strategy.
. The user must be capable of choosing the polling interval
. When the user loads the policy, it must adjust according to what is
the current state of the device.
. The user should not see the policy being execute on system where the
duty cycle is not installed
. The user must be capable of checking what The policy is doing via a log file.
. The user must be capable of configuring the pcb temp sections and
actions on duty cycle.

Change-Id: I1fcb4946f818598fb2406aeb3232af6f4451c018
Signed-off-by: Eduardo Valentin <eduardo.valentin@ti.com>