TIOverlay: Dynamically Switch Framebuffer Color Mode to Fix Alpha
This works with the kernel patch that reverts the default framebuffer
color mode to be RGB24U (Ignore Alpha).

Dynamically switches default framebuffer to RGB8888 on creation
of an overlay, and switches it back to RGB24U on deletion.  This
allows per-pixel alpha when using overlays, but does not
break Composition Bypass optimization for fullscreen 3D GFX
applications.

Change-Id: Ia525dfc428741ded990a2bff4667cbd6a0c91557
Signed-off-by: Chris Tserng <c-tserng@ti.com>