OMAP3: Fix for jerk in Camera Preview
Fix is for jerk / Flicker in Camera Preview, when Camera Application is launched.
Flicker/Jerk in Camera Preview, was observed due to mismatch in buffer which was dequeued
from Camera Sensor and buffer which was send to Display.
Issue is resolved by getting the correct index corresponding to the key, queueToGralloc()
And for that index will get the KeyedVector Gralloc Pointer, which is then send to the
display for preview.

Change-Id: I48642f70de0311d79be0c58eb4aec569193a0bdd
Signed-off-by: Kirti, Badkundri <x0172620@ti.com>