hwc: disable idle timeouts until hwc_set() has been called at least once
The OMAP hwcomposer's uevents listening thread has an idle timeout; if
the timeout (250ms by default) is reached without an event being
recorded, it calls invalidate() to update the screen.  The thread is
started when the hwcomposer module is opened, early in SurfaceFlinger
initialization.   As a result, it's possible that the timeout triggers
before SurfaceFlinger finishes initializing; in that case, the
subsequent call into invalidate() crashes SurfaceFlinger.

Fix this by disabling idle timeouts when the thread starts; this way,
they won't be enabled until hwc_set() has been called, at which point we
know SurfaceFlinger is up and running.

Change-Id: Ic75e7b19efe10b0a20fbc943633ca090006bf5df